requirements.txt
Flask==3.1.0
Pillow==11.0.0
rembg==2.0.60
requests==2.32.3
gunicorn
onnxruntime==1.20.1
